Refactored code so that there can be directories
