+++ doc/TeensyPinouts_sumo_revamp.PNG
