+++ src/device/ProxSense.h
